moved testCheckoutHonorTimeout to GitAPITestUpdateCliGit
